# frontend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Platform - Quant-Service [Development]",
    "description": "Data Service API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/market-data/stock/daily/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "stock"
        ],
        "summary": "일일 주가 데이터 조회",
        "description": "지정된 종목의 일일 주가 데이터(OHLCV)를 조회합니다.",
        "operationId": "Market Data v2-get_daily_prices",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "outputsize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "데이터 크기 (compact: 최근 100일, full: 전체)",
              "default": "compact",
              "title": "Outputsize"
            },
            "description": "데이터 크기 (compact: 최근 100일, full: 전체)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Daily Prices"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/stock/quote/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "stock"
        ],
        "summary": "실시간 주식 호가 조회",
        "description": "지정된 종목의 실시간 호가 정보를 조회합니다.",
        "operationId": "Market Data v2-get_quote",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Quote"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/stock/intraday/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "stock"
        ],
        "summary": "실시간/인트라데이 데이터 조회",
        "description": "지정된 종목의 실시간 또는 분봉 데이터를 조회합니다.",
        "operationId": "Market Data v2-get_intraday_data",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "데이터 간격 (1min, 5min, 15min, 30min, 60min)",
              "default": "1min",
              "title": "Interval"
            },
            "description": "데이터 간격 (1min, 5min, 15min, 30min, 60min)"
          },
          {
            "name": "outputsize",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "데이터 크기 (compact/full)",
              "default": "compact",
              "title": "Outputsize"
            },
            "description": "데이터 크기 (compact/full)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Intraday Data"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/stock/historical/{symbol}": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "stock"
        ],
        "summary": "장기 히스토리 데이터 조회",
        "description": "지정된 종목의 장기 히스토리 데이터를 조회합니다.",
        "operationId": "Market Data v2-get_historical_data",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "종목 심볼 (예: AAPL, TSLA)",
              "title": "Symbol"
            },
            "description": "종목 심볼 (예: AAPL, TSLA)"
          },
          {
            "name": "start_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "시작 날짜 (YYYY-MM-DD)",
              "title": "Start Date"
            },
            "description": "시작 날짜 (YYYY-MM-DD)"
          },
          {
            "name": "end_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "종료 날짜 (YYYY-MM-DD)",
              "title": "End Date"
            },
            "description": "종료 날짜 (YYYY-MM-DD)"
          },
          {
            "name": "frequency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "데이터 주기 (daily, weekly, monthly)",
              "default": "daily",
              "title": "Frequency"
            },
            "description": "데이터 주기 (daily, weekly, monthly)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HistoricalDataResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/stock/symbols": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "stock"
        ],
        "summary": "사용 가능한 종목 심볼 목록 조회",
        "description": "API에서 지원하는 종목 심볼 목록을 조회합니다.",
        "operationId": "Market Data v2-get_available_symbols",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "검색 키워드",
              "title": "Search"
            },
            "description": "검색 키워드"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "결과 개수 제한",
              "default": 100,
              "title": "Limit"
            },
            "description": "결과 개수 제한"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockSymbolsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HistoricalDataResponse": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "data": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Data"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "frequency": {
            "type": "string",
            "title": "Frequency"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "data",
          "count",
          "frequency"
        ],
        "title": "HistoricalDataResponse"
      },
...
      "StockSymbolsResponse": {
        "properties": {
          "symbols": {
            "items": {
              "$ref": "#/components/schemas/SymbolInfo"
            },
            "type": "array",
            "title": "Symbols"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          },
          "search_term": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Search Term"
          }
        },
        "type": "object",
        "required": [
          "symbols",
          "count"
        ],
        "title": "StockSymbolsResponse"
      },
...
      "SymbolInfo": {
        "properties": {
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "region": {
            "type": "string",
            "title": "Region"
          }
        },
        "type": "object",
        "required": [
          "symbol",
          "name",
          "type",
          "region"
        ],
        "title": "SymbolInfo"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}